Fix template for KdTree

+++ include/ugu/accel/kdtree_naive.h
@@ -86,7 +86,7 @@ bool KdTreeNaive<Scalar, Rows>::Build() {
     return false;
   }
   if (Rows < 1) {
-    m_axis_num = m_data[0].rows();
+    m_axis_num = static_cast<int>(m_data[0].rows());
   } else {
     m_axis_num = Rows;
   }
